added test cases for num, str, and if
